+++ src/ragger/backend/ledgerwallet.py
@@ -14,16 +14,15 @@
    limitations under the License.
 """
 from contextlib import contextmanager
-from pathlib import Path
-from typing import Generator, Optional, Any
 from time import sleep
+from typing import Generator, Optional
 
 from ledgerwallet.client import LedgerClient, CommException
 from ledgerwallet.transport import HidDevice
 
-from ragger.utils import RAPDU, Crop
+from ragger.utils import RAPDU
 from ragger.error import ExceptionRAPDU
-from .interface import BackendInterface
+from .physical_backend import PhysicalBackend
 
 
 def raise_policy_enforcer(function):
@@ -45,7 +44,7 @@ def raise_policy_enforcer(function):
     return decoration
 
 
-class LedgerWalletBackend(BackendInterface):
+class LedgerWalletBackend(PhysicalBackend):
 
     def __init__(self, *args, **kwargs) -> None:
         super().__init__(*args, **kwargs)
@@ -101,31 +100,3 @@ class LedgerWalletBackend(BackendInterface):
         self.send_raw(data)
         yield
         self._last_async_response = self.receive()
-
-    def right_click(self) -> None:
-        pass
-
-    def left_click(self) -> None:
-        pass
-
-    def both_click(self) -> None:
-        pass
-
-    def compare_screen_with_snapshot(self,
-                                     golden_snap_path: Path,
-                                     crop: Optional[Crop] = None,
-                                     tmp_snap_path: Optional[Path] = None,
-                                     golden_run: bool = False) -> bool:
-        return True
-
-    def finger_touch(self, x: int = 0, y: int = 0, delay: float = 0.5) -> None:
-        pass
-
-    def wait_for_screen_change(self, timeout: float = 10.0) -> None:
-        return
-
-    def compare_screen_with_text(self, text: str) -> bool:
-        return True
-
-    def get_current_screen_content(self) -> Any:
-        return []