+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- backend: Physical backends (`LedgerComm`, `LedgerWallet`) can now pop a GUI in order to guide a
+           user performing tests on a physical device.
+
 ### Changed
 
 ### Fixed